requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
requests
pillow
langchain-community
sentence-transformers
faiss-cpu